Added report data to logs generated by artillery
Artillery would normally write this log data to a new file which would then be used to generate the html report.  However, AWS Lambda does not support writing files, so I have attached the report data to the logs that are returned from Lambda so that the file can be rebuilt on our end.